HBX-1691: Move class 'org.hibernate.tool.stat.AbstractTreeModel' to 'org.hibernate.tool.internal.util.AbstractTreeModel'
